crypto-portfolio-project/requirements.txt
certifi==2024.6.2
charset-normalizer==3.4.0
idna==3.10
influxdb-client==1.48.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
reactivex==4.0.4
requests==2.32.3
setuptools==75.6.0
six==1.17.0
typing_extensions==4.12.2
urllib3==2.2.3
websockets==14.1
